othia raw, some bugs

+++ anal/skytemp.m
@@ -4,7 +4,7 @@ function t408=skytemp(site,t,az,el)
 global path_GUP
 insert(py.sys.path,int32(0),fullfile(path_GUP,'pygup'))
 getsky=py.importlib.import_module('add_tsky_main');
-t408=single(getsky.matface(fullfile(path_GUP,'matfiles'),site,t,az,el));
+t408=single(getsky.matface(fullfile(path_GUP,'matfiles'),site,int32(t),az,el));
 
 % VHF_30: 150+tsky
 % UHF_ip3: 85+tsky+10*((90-el)*pi/180).^3
